+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.2
+
+* All variables confirmed to lowerCamelCase format in ImgProc & Core.
+* Added HoughCircles as an enchancement feature based on Issue #1.
+* Fixed function names; confirmed to lowerCamelCase
+* Fixed pyrUp, pyrDown, adaptiveThreshold functions
+
 ## 1.0.1
 
 * Fixed initial release bugs/readme bugs.